backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets>=11.0.3,<12.0.0
alpaca-py==0.20.1
pandas==2.1.3
numpy==1.26.2
scikit-learn==1.3.2
xgboost==2.0.2
ta==0.11.0
python-dotenv==1.0.0
httpx==0.25.2
pytz==2023.3.post1
joblib==1.3.2
aiohttp==3.9.1
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
slowapi==0.1.9
sendgrid==6.11.0
